model /requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
python-multipart
Pillow
opencv-python-headless